backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
gunicorn==22.0.0
pandas==2.2.2
psycopg2-binary==2.9.6